Move transpose code to under common
