+++ datafusion/physical-plan/src/common.rs
@@ -171,22 +171,6 @@ pub fn compute_record_batch_statistics(
     }
 }
 
-/// Transposes the given vector of vectors.
-pub fn transpose<T>(original: Vec<Vec<T>>) -> Vec<Vec<T>> {
-    match original.as_slice() {
-        [] => vec![],
-        [first, ..] => {
-            let mut result = (0..first.len()).map(|_| vec![]).collect::<Vec<_>>();
-            for row in original {
-                for (item, transposed_row) in row.into_iter().zip(&mut result) {
-                    transposed_row.push(item);
-                }
-            }
-            result
-        }
-    }
-}
-
 /// Calculates the "meet" of given orderings.
 /// The meet is the finest ordering that satisfied by all the given
 /// orderings, see <https://en.wikipedia.org/wiki/Join_and_meet>.
@@ -701,15 +685,6 @@ mod tests {
         };
 
         assert_eq!(actual, expected);
-        Ok(())
-    }
-
-    #[test]
-    fn test_transpose() -> Result<()> {
-        let in_data = vec![vec![1, 2, 3], vec![4, 5, 6]];
-        let transposed = transpose(in_data);
-        let expected = vec![vec![1, 4], vec![2, 5], vec![3, 6]];
-        assert_eq!(expected, transposed);
         Ok(())
     }
 }
